chore: move UniPoly methods where they should be
